Carve `TestShapeOps` out of the monolith (wala/ML#635).
The 55 shape-manipulation tests (shape reads, reshape, squeeze, split,
slicing, subscripting, indexed layer lists) move verbatim; the monolith
keeps 697, preserving the 1,040 total.
